Adding in UGM library files.
Also adding in mexAllOctave.m, for compiling the library with Octave.  You need
to run either mexAll or mexAllOctave to compile some of the necessary files
before you can run things. Some of the demo display code in examples/ is not
totally compatible with Octave, but it seems like lots of other things are.